Добавил исключения и логирование.

+++ Application/src/main/java/advertisement/services/implementations/AdvertisementService.java
@@ -7,24 +7,31 @@ import advertisement.daos.interfaces.IUserDAO;
 import advertisement.entities.AdvertisementEntity;
 import advertisement.entities.CategoryEntity;
 import advertisement.entities.UserEntity;
+import advertisement.exceptions.invalid.AdvertisementInvalidException;
+import advertisement.exceptions.invalid.FilterInvalidException;
+import advertisement.exceptions.notfound.AdvertisementNotFoundException;
+import advertisement.exceptions.notfound.CategoryNotFoundException;
+import advertisement.exceptions.notfound.UserNotFoundException;
+import advertisement.exceptions.other.AdvertisementIllegalEditException;
 import advertisement.files.interfaces.IFileManager;
 import advertisement.mappers.IAdvertisementModelToEntityMapper;
 import advertisement.models.Advertisement;
 import advertisement.models.Category;
 import advertisement.services.interfaces.IAdvertisementService;
 import jakarta.transaction.Transactional;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
 import java.time.Instant;
-import java.time.LocalDateTime;
 import java.util.*;
-import java.util.stream.Collectors;
 
 @Service
 public class AdvertisementService implements IAdvertisementService {
+    Logger logger = LoggerFactory.getLogger(AdvertisementService.class);
     @Autowired
     private IFileManager fileManager;
     @Autowired
@@ -39,6 +46,12 @@ public class AdvertisementService implements IAdvertisementService {
     @Override
     @Transactional
     public List<Advertisement> getAdvertisements(AdvertisementFilter filter) {
+        logger.info("Получение объявлений.");
+        if (filter == null) {
+            logger.error("Фильтр не указан.");
+            throw new FilterInvalidException("Фильтр не указан.");
+        }
+
         List<Long> categoriesIds = new ArrayList<>();
 
         if (filter.getCategories() != null) {
@@ -53,7 +66,8 @@ public class AdvertisementService implements IAdvertisementService {
                     }
                 }
                 if (!flag) {
-                    throw new NoSuchElementException("Указанная категория не найдена.");
+                    logger.error("Категория {} не найдена.", name);
+                    throw new CategoryNotFoundException("Категория " + name + " не найдена.");
                 }
             }
         }
@@ -65,16 +79,24 @@ public class AdvertisementService implements IAdvertisementService {
                 .comparing(Advertisement::isClosed)
                 .thenComparing(Advertisement::isPaid, Comparator.reverseOrder())
                 .thenComparing(advertisement -> advertisement.getUser().getSellerRating(), Comparator.reverseOrder()));
+
+        logger.info("Объявления успешно получены.");
         return advertisements;
     }
 
     @Override
     @Transactional
     public List<Advertisement> getSalesHistory(String login) {
+        logger.info("Получение истории продаж.");
         Optional<UserEntity> optionalUserEntity = userDAO.findByLogin(login);
-        UserEntity userEntity = optionalUserEntity.orElseThrow();
+        UserEntity userEntity = optionalUserEntity.orElseThrow(() -> {
+            logger.error("Пользователя с таким логином не существует.");
+            throw new UserNotFoundException("Пользователя с таким логином не существует.");
+        });
 
         List<AdvertisementEntity> advertisementEntities = advertisementDAO.findAdvertisementsByUserId(userEntity.getId());
+
+        logger.info("История успешно получена.");
         return advertisementModelToEntityMapper
                 .toModelList(advertisementEntities.stream().filter(AdvertisementEntity::isClosed).toList());
     }
@@ -82,8 +104,21 @@ public class AdvertisementService implements IAdvertisementService {
     @Override
     @Transactional
     public Advertisement addAdvertisement(Advertisement advertisement, MultipartFile multipartFile) throws IOException {
+        logger.info("Добавление объявления.");
+        if (advertisement == null || advertisement.getUser() == null || advertisement.getTitle() == null) {
+            logger.error("Объявление не указано либо не все обязательные поля заполнены.");
+            throw new AdvertisementInvalidException("Объявление не указано либо не все обязательные поля заполнены.");
+        }
+        if (advertisement.getPrice() < 0) {
+            logger.error("Цена товара/услуги не может быть меньше 0.");
+            throw new AdvertisementInvalidException("Цена товара/услуги не может быть меньше 0.");
+        }
+
         Optional<UserEntity> optionalUserEntity = userDAO.findByLogin(advertisement.getUser().getLogin());
-        UserEntity userEntity = optionalUserEntity.orElseThrow();
+        UserEntity userEntity = optionalUserEntity.orElseThrow(() -> {
+            logger.error("Пользователя с таким логином не существует.");
+            throw new UserNotFoundException("Пользователя с таким логином не существует.");
+        });
 
         if (multipartFile != null) {
             String previewLink = fileManager.savePreview(multipartFile);
@@ -103,7 +138,8 @@ public class AdvertisementService implements IAdvertisementService {
                 }
             }
             if (!flag) {
-                throw new NoSuchElementException("Указанная категория не найдена.");
+                logger.error("Категория {} не найдена.", category.getName());
+                throw new CategoryNotFoundException("Категория " + category.getName() + " не найдена.");
             }
         }
 
@@ -111,14 +147,28 @@ public class AdvertisementService implements IAdvertisementService {
         advertisementEntity.setPublished(instant);
         advertisementEntity.setUser(userEntity);
 
+        logger.info("Объявление успешно добавлено.");
         return advertisementModelToEntityMapper.toModel(advertisementDAO.save(advertisementEntity));
     }
 
     @Override
     @Transactional
-    public Advertisement editAdvertisement(Advertisement model, MultipartFile multipartFile) throws IOException, IllegalAccessException {
+    public Advertisement editAdvertisement(Advertisement model, MultipartFile multipartFile) throws IOException {
+        logger.info("Редактирование объявления.");
+        if (model == null || model.getUser() == null || model.getTitle() == null) {
+            logger.error("Объявление не указано либо не все обязательные поля заполнены.");
+            throw new AdvertisementInvalidException("Объявление не указано либо не все обязательные поля заполнены.");
+        }
+        if (model.getPrice() < 0) {
+            logger.error("Цена товара/услуги не может быть меньше 0.");
+            throw new AdvertisementInvalidException("Цена товара/услуги не может быть меньше 0.");
+        }
+
         Optional<AdvertisementEntity> optionalAdvertisementEntity = advertisementDAO.findByAdNumber(model.getAdNumber());
-        AdvertisementEntity advertisementEntity = optionalAdvertisementEntity.orElseThrow();
+        AdvertisementEntity advertisementEntity = optionalAdvertisementEntity.orElseThrow(() -> {
+            logger.error("Объявления с таким артикулом не существует.");
+            throw new AdvertisementNotFoundException("Объявления с таким артикулом не существует.");
+        });
         advertisementEntity.setCategories(new ArrayList<>());
 
         List<Advertisement> advertisements = advertisementModelToEntityMapper
@@ -130,7 +180,8 @@ public class AdvertisementService implements IAdvertisementService {
                         Objects.equals(advertisement.getUser().getLogin(), model.getUser().getLogin()));
 
         if (!isOwner) {
-            throw new IllegalAccessException("Редактировать объявление может только его владелец.");
+            logger.error("Редактировать объявление может только его владелец.");
+            throw new AdvertisementIllegalEditException("Редактировать объявление может только его владелец.");
         }
 
         List<CategoryEntity> categoryEntities = categoryDAO.findAll();
@@ -145,7 +196,8 @@ public class AdvertisementService implements IAdvertisementService {
                 }
             }
             if (!flag) {
-                throw new NoSuchElementException("Указанная категория не найдена.");
+                logger.error("Категория {} не найдена.", category.getName());
+                throw new CategoryNotFoundException("Категория " + category.getName() + " не найдена.");
             }
         }
 
@@ -163,24 +215,46 @@ public class AdvertisementService implements IAdvertisementService {
         advertisementEntity.setTown(model.getTown());
 
         advertisementDAO.update(advertisementEntity);
+
+        logger.info("Объявление успешно отредактировано.");
         return advertisementModelToEntityMapper.toModel(advertisementEntity);
     }
 
     @Override
     @Transactional
     public Advertisement prepayAdvertisement(Advertisement model) {
+        logger.info("Проплата объявления.");
+        if (model == null) {
+            logger.error("Объявление не указано.");
+            throw new AdvertisementInvalidException("Объявление не указано.");
+        }
+
         Optional<AdvertisementEntity> optionalAdvertisementEntity = advertisementDAO.findByAdNumber(model.getAdNumber());
-        AdvertisementEntity advertisementEntity = optionalAdvertisementEntity.orElseThrow();
+        AdvertisementEntity advertisementEntity = optionalAdvertisementEntity.orElseThrow(() -> {
+            logger.error("Объявления с таким артикулом не существует.");
+            throw new AdvertisementNotFoundException("Объявления с таким артикулом не существует.");
+        });
         advertisementEntity.setPaid(true);
         advertisementDAO.update(advertisementEntity);
+
+        logger.info("Объявление успешно проплачено.");
         return advertisementModelToEntityMapper.toModel(advertisementEntity);
     }
 
     @Override
     @Transactional
-    public Advertisement closeAdvertisement(Advertisement model) throws IllegalAccessException {
+    public Advertisement closeAdvertisement(Advertisement model) {
+        logger.info("Закрытие объявления.");
+        if (model == null || model.getUser() == null) {
+            logger.error("Объявление не указано либо не все обязательные поля заполнены.");
+            throw new AdvertisementInvalidException("Объявление не указано либо не все обязательные поля заполнены.");
+        }
+
         Optional<AdvertisementEntity> optionalAdvertisementEntity = advertisementDAO.findByAdNumber(model.getAdNumber());
-        AdvertisementEntity advertisementEntity = optionalAdvertisementEntity.orElseThrow();
+        AdvertisementEntity advertisementEntity = optionalAdvertisementEntity.orElseThrow(() -> {
+            logger.error("Объявления с таким артикулом не существует.");
+            throw new AdvertisementNotFoundException("Объявления с таким артикулом не существует.");
+        });
         List<Advertisement> advertisements = advertisementModelToEntityMapper
                 .toModelList(advertisementDAO
                         .findAdvertisementsByUserId(advertisementEntity.getUser().getId()));
@@ -190,22 +264,35 @@ public class AdvertisementService implements IAdvertisementService {
                         Objects.equals(advertisement.getUser().getLogin(), model.getUser().getLogin()));
 
         if (!isOwner) {
-            throw new IllegalAccessException("Закрыть объявление может только его владелец.");
+            logger.error("Редактировать объявление может только его владелец.");
+            throw new AdvertisementIllegalEditException("Редактировать объявление может только его владелец.");
         }
 
         advertisementEntity.setClosed(true);
         advertisementDAO.update(advertisementEntity);
+
+        logger.info("Объяаление успешно закрыто.");
         return advertisementModelToEntityMapper.toModel(advertisementEntity);
     }
 
     @Override
     @Transactional
     public Advertisement deleteAdvertisement(Advertisement model) throws IOException {
+        logger.info("Удаление объявления.");
+        if (model == null) {
+            logger.error("Объявление не указано.");
+            throw new AdvertisementInvalidException("Объявление не указано.");
+        }
+
         Optional<AdvertisementEntity> optionalAdvertisementEntity = advertisementDAO.findByAdNumber(model.getAdNumber());
-        AdvertisementEntity advertisementEntity = optionalAdvertisementEntity.orElseThrow();
+        AdvertisementEntity advertisementEntity = optionalAdvertisementEntity.orElseThrow(() -> {
+            logger.error("Объявления с таким артикулом не существует.");
+            throw new AdvertisementNotFoundException("Объявления с таким артикулом не существует.");
+        });
         advertisementDAO.delete(advertisementEntity);
         fileManager.deleteOldPreview(advertisementEntity.getPreviewLink());
+
+        logger.info("Объявление успешно удалено.");
         return advertisementModelToEntityMapper.toModel(advertisementEntity);
     }
-
 }
